Extract ManagedTextSection
