--- src/Compilers/Core/Portable/System.Reflection.Metadata/PortableExecutable/PESizes.cs
@@ -1,48 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-namespace System.Reflection.PortableExecutable
-{
-    internal sealed class PESizes
-    {
-        /// <summary>
-        /// Total size of metadata (header and all streams).
-        /// </summary>
-        public readonly int MetadataSize;
-
-        /// <summary>
-        /// The size of IL stream.
-        /// </summary>
-        public readonly int ILStreamSize;
-
-        /// <summary>
-        /// The size of mapped field data stream.
-        /// Aligned to <see cref="Microsoft.Cci.MetadataWriter.MappedFieldDataAlignment"/>.
-        /// </summary>
-        public readonly int MappedFieldDataSize;
-
-        /// <summary>
-        /// The size of managed resource data stream.
-        /// Aligned to <see cref="Microsoft.Cci.MetadataWriter.ManagedResourcesDataAlignment"/>.
-        /// </summary>
-        public readonly int ResourceDataSize;
-
-        /// <summary>
-        /// Size of strong name hash.
-        /// </summary>
-        public readonly int StrongNameSignatureSize;
-
-        public PESizes(
-            int metadataSize,
-            int ilStreamSize,
-            int mappedFieldDataSize,
-            int resourceDataSize,
-            int strongNameSignatureSize)
-        {
-            MetadataSize = metadataSize;
-            ResourceDataSize = resourceDataSize;
-            ILStreamSize = ilStreamSize;
-            MappedFieldDataSize = mappedFieldDataSize;
-            StrongNameSignatureSize = strongNameSignatureSize;
-        }
-    }
-}
